requirements.txt
fastapi>=0.115.0
uvicorn[standard]==0.30.0

sqlalchemy==2.0.30
psycopg[binary]

pydantic>=2.9.2
python-dotenv==1.0.1

alembic==1.13.1

passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

email-validator==2.1.1